gh-2650 Refactor TestExpressionParser into multiple classes
